# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /order:
    post:
      operationId: createPaymentOrder
      tags: [bookings]
      summary: Create a payment order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentOrderInput"
      responses:
        "200":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
        "400":
          description: Invalid payment order request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Payment provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verify:
    post:
      operationId: verifyPayment
      tags: [bookings]
      summary: Verify a payment and create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentInput"
      responses:
        "201":
          description: Booking created after payment verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid payment or booking request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Coach:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        activity:
          type: string
        venue:
          type: string
        area:
          type: string
        distance:
          type: string
        rating:
          type: number
        reviews:
          type: integer
        price:
          type: integer
        priceLabel:
          type: string
        ageRange:
          type: string
        experience:
          type: string
        description:
          type: string
        verified:
          type: boolean
        trialAvailable:
          type: boolean
        imageUrl:
          type: string
        accent:
          type: string
        highlights:
          type: array
          items:
            type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TrialSlot"
        sessionFormats:
          type: array
          items:
            type: string
        venueType:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        parentAccompanied:
          type: boolean
      required:
        - id
        - name
        - activity
        - venue
        - area
        - distance
        - rating
        - reviews
        - price
        - priceLabel
        - ageRange
        - experience
        - description
        - verified
        - trialAvailable
        - imageUrl
        - accent
        - highlights
        - slots
        - sessionFormats
        - venueType
        - serviceAreas
        - ageMin
        - ageMax
        - parentAccompanied
    TrialSlot:
      type: object
      properties:
        id:
          type: string
        day:
          type: string
        date:
          type: string
        time:
          type: string
        label:
          type: string
      required:
        - id
        - day
        - date
        - time
        - label
    BookingInput:
      type: object
      properties:
        coachId:
          type: string
        slotId:
          type: string
        childName:
          type: string
        childAge:
          type: integer
          minimum: 4
          maximum: 18
        parentName:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        seats:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - coachId
        - slotId
        - childName
        - childAge
        - parentName
        - parentEmail
        - parentPhone
    PaymentOrderInput:
      type: object
      properties:
        coachId:
          type: string
          minLength: 1
        slotId:
          type: string
          minLength: 1
        seats:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - coachId
        - slotId
        - seats
    PaymentOrder:
      type: object
      properties:
        mode:
          type: string
          enum: [razorpay, simulated]
        orderId:
          type: string
        keyId:
          type: string
        amount:
          type: integer
        amountPaise:
          type: integer
        currency:
          type: string
          enum: [INR]
        coachFee:
          type: integer
        dabbleFee:
          type: integer
        total:
          type: integer
        seats:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - mode
        - orderId
        - keyId
        - amount
        - amountPaise
        - currency
        - coachFee
        - dabbleFee
        - total
        - seats
    VerifyPaymentInput:
      allOf:
        - $ref: "#/components/schemas/BookingInput"
        - type: object
          properties:
            orderId:
              type: string
              minLength: 1
            paymentId:
              type: string
            signature:
              type: string
          required:
            - orderId
            - seats
    Booking:
      allOf:
        - $ref: "#/components/schemas/BookingInput"
        - type: object
          properties:
            id:
              type: string
            coach:
              $ref: "#/components/schemas/Coach"
            slot:
              $ref: "#/components/schemas/TrialSlot"
            trialFee:
              type: integer
            serviceFee:
              type: integer
            coachFee:
              type: integer
            dabbleFee:
              type: integer
            seats:
              type: integer
              minimum: 1
              maximum: 10
            total:
              type: integer
            createdAt:
              type: string
          required:
            - id
            - coach
            - slot
            - trialFee
            - serviceFee
             - coachFee
             - dabbleFee
             - seats
            - total
            - createdAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...